+++ src/modules/database/types/database-cluster-mysql-user-settings.ts
@@ -1,0 +1,3 @@
+export interface IDatabaseClusterMysqlUserSettings {
+  auth_plugin: string;
+}