+++ src/modules/droplet/create-droplets/create-droplets.mock.ts
@@ -19,6 +19,7 @@ export const request = {
     "tags": ["web"],
     "user_data": 'test-user-data',
     "volumes": [],
+    "vpc_uuid": "vpc_uuid",
   },
   "minimumBody": {
     "names": [