+++ src/modules/kubernetes/types/index.ts
@@ -1,4 +1,5 @@
 export * from './kubernetes-cluster-maintenance-policy';
+export * from './kubernetes-cluster-node-pool-labels';
 export * from './kubernetes-cluster-node-pool-node-status';
 export * from './kubernetes-cluster-node-pool-node';
 export * from './kubernetes-cluster-node-pool';