+++ src/modules/kubernetes/types/kubernetes-cluster-node-pool-labels.ts
@@ -1,0 +1,3 @@
+export interface IKubernetesClusterNodePoolLabels {
+  [key: string]: string;
+}