+++ src/modules/load-balancer/create-load-balancer/create-load-balancer.ts
@@ -1,8 +1,22 @@
 import { IResponse, IContext } from '../../../types';
-import { ILoadBalancer } from '..';
+import { ILoadBalancer, IForwardingRule, IHealthCheck, IStickSessions } from '..';
 
 export interface ICreateLoadBalancerApiResponse {
   load_balancer: ILoadBalancer;
+}
+
+export interface ICreateLoadBalancerApiRequest {
+  algorithm?: string;
+  droplet_ids?: number[];
+  enable_proxy_protocol?: boolean;
+  forwarding_rules: IForwardingRule[];
+  health_check?: IHealthCheck;
+  name: string;
+  redirect_http_to_https?: boolean;
+  region: string;
+  sticky_sessions?: IStickSessions;
+  tag?: string;
+  vpc_uuid?: string;
 }
 
 export type CreateLoadBalancerResponse = IResponse<ICreateLoadBalancerApiResponse>;
@@ -19,7 +33,9 @@ export const createLoadBalancer = ({
   redirect_http_to_https,
   region,
   sticky_sessions,
-}: ILoadBalancer): Promise<Readonly<CreateLoadBalancerResponse>> => {
+  tag,
+  vpc_uuid,
+}: ICreateLoadBalancerApiRequest): Promise<Readonly<CreateLoadBalancerResponse>> => {
   const path = '/load_balancers';
   const body = {
     algorithm,
@@ -31,6 +47,8 @@ export const createLoadBalancer = ({
     redirect_http_to_https,
     region,
     sticky_sessions,
+    tag,
+    vpc_uuid,
   };
   const url = `${path}`;