+++ src/modules/vpc/index.ts
@@ -1,0 +1,7 @@
+export * from './create-vpc/create-vpc';
+export * from './delete-vpc/delete-vpc';
+export * from './get-vpc/get-vpc';
+export * from './list-vpc-resources/list-vpc-resources';
+export * from './list-vpcs/list-vpcs';
+export * from './types';
+export * from './update-vpc/update-vpc';