+++ src/modules/vpc/list-vpcs/list-vpcs.ts
@@ -1,0 +1,21 @@
+import { IResponse, IContext, IListResponse, IListRequest } from '../../../types';
+import { IVpc } from '..';
+
+export interface IListVpcsApiResponse extends IListResponse {
+  vpcs: IVpc[];
+}
+
+export type ListVpcsResponse = IResponse<IListVpcsApiResponse>;
+
+export const listVpcs = ({
+  httpClient,
+}: IContext) => ({
+  page = 1,
+  per_page = 25,
+}: IListRequest): Promise<Readonly<ListVpcsResponse>> => {
+  const path = '/vpcs';
+  const queryParams = {page, per_page};
+  const url = `${path}`;
+
+  return httpClient.get<IListVpcsApiResponse>(url, {params: queryParams});
+};