+++ src/modules/vpc/types/index.ts
@@ -1,0 +1,2 @@
+export * from './vpc';
+export * from './vpc-resource';