+++ src/modules/vpc/types/vpc-resource.ts
@@ -1,0 +1,5 @@
+export interface IVpcResource {
+  created_at: string;
+  name: string;
+  urn: string;
+}